+++ app/src/main/java/eu/faircode/email/DB.java
@@ -67,7 +67,7 @@ import javax.mail.internet.InternetAddress;
 // https://developer.android.com/topic/libraries/architecture/room.html
 
 @Database(
-        version = 258,
+        version = 260,
         entities = {
                 EntityIdentity.class,
                 EntityAccount.class,
@@ -2609,11 +2609,18 @@ public abstract class DB extends RoomDatabase {
                         logMigration(startVersion, endVersion);
                         defaultSearches(db, context);
                     }
-                }).addMigrations(new Migration(258, 257) {
+                }).addMigrations(new Migration(258, 259) {
                     @Override
                     public void migrate(@NonNull SupportSQLiteDatabase db) {
                         logMigration(startVersion, endVersion);
-                        defaultSearches(db, context);
+                        db.execSQL("UPDATE account SET keep_alive_noop = 0" +
+                                " WHERE host = 'outlook.office365.com' AND pop = " + EntityAccount.TYPE_IMAP);
+                    }
+                }).addMigrations(new Migration(259, 260) {
+                    @Override
+                    public void migrate(@NonNull SupportSQLiteDatabase db) {
+                        logMigration(startVersion, endVersion);
+                        db.execSQL("ALTER TABLE `rule` ADD COLUMN `daily` INTEGER NOT NULL DEFAULT 0");
                     }
                 }).addMigrations(new Migration(998, 999) {
                     @Override