+++ app/src/main/java/eu/faircode/email/DeepL.java
@@ -70,6 +70,7 @@ import javax.net.ssl.HttpsURLConnection;
 
 public class DeepL {
     // https://www.deepl.com/docs-api/
+    // https://github.com/DeepLcom/deepl-java
     private static JSONArray jlanguages = null;
 
     private static final int DEEPL_TIMEOUT = 20; // seconds
@@ -424,7 +425,7 @@ public class DeepL {
                             next.set(Calendar.MILLISECOND, 0);
                             next.set(Calendar.SECOND, 0);
                             next.set(Calendar.MINUTE, 0);
-                            next.set(Calendar.HOUR, 0);
+                            next.set(Calendar.HOUR_OF_DAY, 0);
                             long today = next.getTimeInMillis();
                             if (next.get(Calendar.DATE) > subscription)
                                 next.add(Calendar.MONTH, 1);