+++ app/src/main/java/eu/faircode/email/FragmentRule.java
@@ -88,6 +88,8 @@ public class FragmentRule extends FragmentBase {
     private EditText etName;
     private EditText etOrder;
     private CheckBox cbEnabled;
+    private CheckBox cbDaily;
+    private EditText etAge;
     private CheckBox cbStop;
 
     private EditText etSender;
@@ -165,6 +167,7 @@ public class FragmentRule extends FragmentBase {
     private ContentLoadingProgressBar pbWait;
 
     private Group grpReady;
+    private Group grpAge;
     private Group grpSnooze;
     private Group grpFlag;
     private Group grpImportance;
@@ -244,6 +247,8 @@ public class FragmentRule extends FragmentBase {
         etName = view.findViewById(R.id.etName);
         etOrder = view.findViewById(R.id.etOrder);
         cbEnabled = view.findViewById(R.id.cbEnabled);
+        cbDaily = view.findViewById(R.id.cbDaily);
+        etAge = view.findViewById(R.id.etAge);
         cbStop = view.findViewById(R.id.cbStop);
 
         etSender = view.findViewById(R.id.etSender);
@@ -322,6 +327,7 @@ public class FragmentRule extends FragmentBase {
         pbWait = view.findViewById(R.id.pbWait);
 
         grpReady = view.findViewById(R.id.grpReady);
+        grpAge = view.findViewById(R.id.grpAge);
         grpSnooze = view.findViewById(R.id.grpSnooze);
         grpFlag = view.findViewById(R.id.grpFlag);
         grpImportance = view.findViewById(R.id.grpImportance);
@@ -334,6 +340,13 @@ public class FragmentRule extends FragmentBase {
         grpAutomation = view.findViewById(R.id.grpAutomation);
         grpDelete = view.findViewById(R.id.grpDelete);
         grpLocalOnly = view.findViewById(R.id.grpLocalOnly);
+
+        cbDaily.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                grpAge.setVisibility(isChecked ? View.VISIBLE : View.GONE);
+            }
+        });
 
         ibSender.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -732,6 +745,7 @@ public class FragmentRule extends FragmentBase {
         tvFolder.setText(null);
         bottom_navigation.setVisibility(View.GONE);
         grpReady.setVisibility(View.GONE);
+        grpAge.setVisibility(View.GONE);
         grpSnooze.setVisibility(View.GONE);
         grpFlag.setVisibility(View.GONE);
         grpImportance.setVisibility(View.GONE);
@@ -1087,6 +1101,7 @@ public class FragmentRule extends FragmentBase {
                         JSONObject jcondition = (rule == null ? new JSONObject() : new JSONObject(rule.condition));
                         JSONObject jaction = (rule == null ? new JSONObject() : new JSONObject(rule.action));
 
+                        JSONObject jgeneral = jcondition.optJSONObject("general");
                         JSONObject jsender = jcondition.optJSONObject("sender");
                         JSONObject jrecipient = jcondition.optJSONObject("recipient");
                         JSONObject jsubject = jcondition.optJSONObject("subject");
@@ -1098,6 +1113,8 @@ public class FragmentRule extends FragmentBase {
                         etName.setText(rule == null ? args.getString("subject") : rule.name);
                         etOrder.setText(rule == null ? null : Integer.toString(rule.order));
                         cbEnabled.setChecked(rule == null || rule.enabled);
+                        cbDaily.setChecked(rule != null && rule.daily);
+                        etAge.setText(jgeneral == null ? null : Integer.toString(jgeneral.optInt("age")));
                         cbStop.setChecked(rule != null && rule.stop);
 
                         etSender.setText(jsender == null ? args.getString("sender") : jsender.getString("value"));
@@ -1249,6 +1266,7 @@ public class FragmentRule extends FragmentBase {
                     Log.e(ex);
                 } finally {
                     grpReady.setVisibility(View.VISIBLE);
+                    grpAge.setVisibility(cbDaily.isChecked() ? View.VISIBLE : View.GONE);
                     if (id < 0)
                         bottom_navigation.getMenu().removeItem(R.id.action_delete);
                     bottom_navigation.setVisibility(View.VISIBLE);
@@ -1295,6 +1313,7 @@ public class FragmentRule extends FragmentBase {
 
             Bundle args = new Bundle();
             args.putLong("folder", folder);
+            args.putBoolean("daily", cbDaily.isChecked());
             args.putString("condition", jcondition.toString());
             args.putString("action", jaction.toString());
 
@@ -1320,6 +1339,7 @@ public class FragmentRule extends FragmentBase {
             args.putString("name", etName.getText().toString());
             args.putString("order", etOrder.getText().toString());
             args.putBoolean("enabled", cbEnabled.isChecked());
+            args.putBoolean("daily", cbDaily.isChecked());
             args.putBoolean("stop", cbStop.isChecked());
             args.putString("condition", getCondition().toString());
             args.putString("action", getAction().toString());
@@ -1342,6 +1362,7 @@ public class FragmentRule extends FragmentBase {
                     String name = args.getString("name");
                     String order = args.getString("order");
                     boolean enabled = args.getBoolean("enabled");
+                    boolean daily = args.getBoolean("daily");
                     boolean stop = args.getBoolean("stop");
                     String condition = args.getString("condition");
                     String action = args.getString("action");
@@ -1378,6 +1399,7 @@ public class FragmentRule extends FragmentBase {
                         rule.name = name;
                         rule.order = Integer.parseInt(order);
                         rule.enabled = enabled;
+                        rule.daily = daily;
                         rule.stop = stop;
                         rule.condition = condition;
                         rule.action = action;
@@ -1389,6 +1411,7 @@ public class FragmentRule extends FragmentBase {
                         rule.name = name;
                         rule.order = Integer.parseInt(order);
                         rule.enabled = enabled;
+                        rule.daily = daily;
                         rule.stop = stop;
                         rule.condition = condition;
                         rule.action = action;
@@ -1420,6 +1443,18 @@ public class FragmentRule extends FragmentBase {
 
     private JSONObject getCondition() throws JSONException {
         JSONObject jcondition = new JSONObject();
+
+        JSONObject jgeneral = new JSONObject();
+
+        String age = etAge.getText().toString().trim();
+        if (!TextUtils.isEmpty(age) && TextUtils.isDigitsOnly(age))
+            try {
+                jgeneral.put("age", Integer.parseInt(age));
+            } catch (Throwable ex) {
+                Log.e(ex);
+            }
+
+        jcondition.put("general", jgeneral);
 
         String sender = etSender.getText().toString();
         boolean known = cbKnownSender.isChecked();
@@ -1635,6 +1670,7 @@ public class FragmentRule extends FragmentBase {
         @Override
         public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
             long folder = getArguments().getLong("folder");
+            boolean daily = getArguments().getBoolean("daily");
             String condition = getArguments().getString("condition");
             String action = getArguments().getString("action");
 
@@ -1657,6 +1693,7 @@ public class FragmentRule extends FragmentBase {
 
             final Bundle args = new Bundle();
             args.putLong("folder", folder);
+            args.putBoolean("daily", daily);
             args.putString("condition", condition);
             args.putString("action", action);
 
@@ -1682,6 +1719,7 @@ public class FragmentRule extends FragmentBase {
                         protected Integer onExecute(Context context, Bundle args) throws Throwable {
                             EntityRule rule = new EntityRule();
                             rule.folder = args.getLong("folder");
+                            rule.daily = args.getBoolean("daily");
                             rule.condition = args.getString("condition");
                             rule.action = args.getString("action");
 
@@ -1695,7 +1733,7 @@ public class FragmentRule extends FragmentBase {
                                     db.beginTransaction();
 
                                     EntityMessage message = db.message().getMessage(mid);
-                                    if (message == null)
+                                    if (message == null || message.ui_hide)
                                         continue;
 
                                     if (rule.matches(context, message, null, null))
@@ -1745,6 +1783,7 @@ public class FragmentRule extends FragmentBase {
                 protected List<EntityMessage> onExecute(Context context, Bundle args) throws Throwable {
                     EntityRule rule = new EntityRule();
                     rule.folder = args.getLong("folder");
+                    rule.daily = args.getBoolean("daily");
                     rule.condition = args.getString("condition");
                     rule.action = args.getString("action");
                     rule.validate(context);