+++ app/src/main/java/eu/faircode/email/MessageHelper.java
@@ -239,7 +239,9 @@ public class MessageHelper {
             "$HasAttachment", // Dovecot
             "$HasNoAttachment", // Dovecot
             "$IsTrusted", // Fastmail
-            "$X-ME-Annot-2" // Fastmail
+            "$X-ME-Annot-2", // Fastmail
+            "$purchases", // mailbox.org
+            "$social " // mailbox.org
     ));
 
     // https://tools.ietf.org/html/rfc4021
@@ -962,16 +964,27 @@ public class MessageHelper {
                     HtmlHelper.autoLink(document, true);
                 }
 
-                StringBuilder style = new StringBuilder();
+                if (!TextUtils.isEmpty(compose_font) || compose_color != Color.TRANSPARENT) {
+                    List<Node> childs = new ArrayList<>();
+                    for (Node child : document.body().childNodes())
+                        if (TextUtils.isEmpty(child.attr("fairemail"))) {
+                            childs.add(child);
+                            child.remove();
+                        } else
+                            break;
 
-                if (compose_color != Color.TRANSPARENT)
-                    style.append("* {color: ").append(HtmlHelper.encodeWebColor(compose_color)).append(";}").append('\n');
+                    StringBuilder style = new StringBuilder();
+                    if (!TextUtils.isEmpty(compose_font))
+                        style.append("font-family: ").append(StyleHelper.getFamily(compose_font)).append(';');
+                    if (compose_color != Color.TRANSPARENT)
+                        style.append("color: ").append(HtmlHelper.encodeWebColor(compose_color)).append(';');
 
-                if (!TextUtils.isEmpty(compose_font))
-                    style.append("* {font-family: ").append(StyleHelper.getFamily(compose_font)).append(";}").append('\n');
+                    Element div = document.createElement("div").attr("style", style.toString());
 
-                if (style.length() > 0)
-                    document.head().append("<style>" + style + "</style>");
+                    for (Node child : childs)
+                        div.appendChild(child);
+                    document.body().prependChild(div);
+                }
 
                 document.select("div[fairemail=signature]").removeAttr("fairemail");
                 document.select("div[fairemail=reference]").removeAttr("fairemail");