+++ CHANGELOG.md
@@ -6,6 +6,30 @@
 
 For support you can use [the contact form](https://contact.faircode.eu/?product=fairemailsupport).
 
+### 1.2017 - 2022-12-23
+
+* Added option to disable unread shadow border
+* Small improvements and minor bug fixes
+* Updated translations
+
+### 1.2016 - 2022-12-18
+
+* Fixed crash introduced in previous version
+* Small improvements and minor bug fixes
+
+### 1.2015 - 2022-12-18
+
+* Fixed default compose font / text color
+* Updated translations
+
+### 1.2014 - 2022-12-17
+
+* Added saved search for specific account/folder
+* Added option for default text color for message composer
+* Added option to swap trash/archive button in bottom action bar
+* Small improvements and minor bug fixes
+* Updated translations
+
 ### 1.2013 - 2022-12-09
 
 * Added scrolling/fixed style toolbar